Application Class を追加
- なぜかNavigation Actions がApplicationファイルに含まれていたためNavGraphへ移動
- Application Class を追加
- Application Class に@ HiltAndroidApp を追加
- Manifest に追加したApplicationクラスを追加
