+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,8 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.TwitterTraceAndroid">
+        android:theme="@style/Theme.TwitterTraceAndroid"
+        android:name=".TwitterTraceApplication">
         <activity
             android:name=".TestActivity"
             android:exported="false"
