+++ app/src/main/java/com/example/twitter_trace_android/ui/TwitterTraceNavGraph.kt
@@ -1,20 +1,16 @@
 package com.example.twitter_trace_android.ui
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.twitter_trace_android.TwitterTraceDestinations
 import com.example.twitter_trace_android.data.repository.tweet.impl.FakeTweetRepository
 import com.example.twitter_trace_android.data.repository.user.impl.FakeUserRepository
 import com.example.twitter_trace_android.ui.timeline.TimelineRoute
-import com.example.twitter_trace_android.ui.timeline.TimelineScreen
-import com.example.twitter_trace_android.ui.timeline.TimelineUiState
 import com.example.twitter_trace_android.ui.timeline.TimelineViewModel
 
 @Composable
@@ -35,10 +31,34 @@ fun TwitterTraceNavGraph(
                     FakeTweetRepository()
                 )
             )
+//            val timelineViewModel:TimelineViewMode by viewModels()
 
             TimelineRoute(
                 viewModel = timelineViewModel
             )
         }
     }
+}
+
+
+/**
+ */
+object TwitterTraceDestinations {
+    const val TIMELINE_ROUTE = "timeline"
+}
+
+class TwitterTraceActions(navController: NavHostController) {
+    val navigateToTimeline: () -> Unit = {
+        navController.navigate(TwitterTraceDestinations.TIMELINE_ROUTE) {
+            // ユーザーがアイテムを選択するときに、バックスタックに宛先の大きなスタックが構築されないように、グラフの開始宛先までポップアップします。
+            popUpTo(navController.graph.findStartDestination().id) {
+                saveState = true
+            }
+            // 同じアイテムを再選択するときは、同じ宛先の複数のコピーを避けてください
+            launchSingleTop = true
+            // 以前に選択したアイテムを再選択すると、状態が復元されます
+            restoreState = true
+        }
+    }
+
 }
